SSE onDone cleanup fix & moved SSEDisplay to its own file
