+++ lib/widgets/response_body_success.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 
+import 'package:apidash/widgets/sse_display.dart';
 import 'package:apidash_core/apidash_core.dart';
 import 'package:apidash_design_system/apidash_design_system.dart';
 import 'package:flutter/foundation.dart';
@@ -161,137 +162,6 @@ class _ResponseBodySuccessState extends State<ResponseBodySuccess> {
           ),
         );
       },
-    );
-  }
-}
-
-//MOVE THIS SOMEWHERE ELSE
-class SSEDisplay extends StatefulWidget {
-  final String sseOutput;
-  const SSEDisplay({super.key, required this.sseOutput});
-
-  @override
-  State<SSEDisplay> createState() => _SSEDisplayState();
-}
-
-class _SSEDisplayState extends State<SSEDisplay> {
-  final _scrollController = ScrollController();
-  bool autoScrollEnabled = true;
-  bool _isScrolling = false;
-
-  @override
-  void initState() {
-    super.initState();
-
-    _scrollController.addListener(() {
-      final position = _scrollController.position;
-      final atBottom = position.pixels >= position.maxScrollExtent - 50;
-
-      if (autoScrollEnabled && !atBottom) {
-        // User scrolled up manually
-        setState(() => autoScrollEnabled = false);
-      } else if (!autoScrollEnabled && atBottom) {
-        // User scrolled back to bottom
-        setState(() => autoScrollEnabled = true);
-      }
-    });
-  }
-
-  @override
-  void didUpdateWidget(covariant SSEDisplay oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    if (oldWidget.sseOutput != widget.sseOutput &&
-        autoScrollEnabled &&
-        !_isScrolling) {
-      WidgetsBinding.instance.addPostFrameCallback((_) async {
-        if (_scrollController.hasClients) {
-          _isScrolling = true;
-          _scrollController.jumpTo(
-            _scrollController.position.maxScrollExtent,
-          );
-          _isScrolling = false;
-        }
-      });
-    }
-  }
-
-  @override
-  void dispose() {
-    _scrollController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    List<dynamic> sse;
-    try {
-      sse = jsonDecode(widget.sseOutput);
-    } catch (e) {
-      return Text(
-        'Invalid SSE output',
-        style: theme.textTheme.bodyMedium?.copyWith(color: Colors.red),
-      );
-    }
-
-    return SingleChildScrollView(
-      controller: _scrollController,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: sse.map<Widget>((chunk) {
-          Map<String, dynamic>? parsedJson;
-          try {
-            parsedJson = jsonDecode(chunk);
-          } catch (_) {}
-
-          return Card(
-            margin: const EdgeInsets.symmetric(vertical: 6, horizontal: 8),
-            elevation: 2,
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(12),
-            ),
-            child: Padding(
-              padding: const EdgeInsets.all(12.0),
-              child: parsedJson != null
-                  ? Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: parsedJson.entries.map((entry) {
-                        return Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 2.0),
-                          child: Row(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Text(
-                                '${entry.key}: ',
-                                style: theme.textTheme.bodyMedium?.copyWith(
-                                  fontWeight: FontWeight.bold,
-                                  color: kColorGQL,
-                                ),
-                              ),
-                              const SizedBox(width: 4),
-                              Expanded(
-                                child: Text(
-                                  entry.value.toString(),
-                                  style: theme.textTheme.bodyMedium?.copyWith(
-                                    fontFamily: 'monospace',
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        );
-                      }).toList(),
-                    )
-                  : Text(
-                      chunk.toString(),
-                      style: theme.textTheme.bodyMedium?.copyWith(
-                        fontFamily: 'monospace',
-                      ),
-                    ),
-            ),
-          );
-        }).toList(),
-      ),
     );
   }
 }
